+++ src/instance/methods/sync.js
@@ -14,6 +14,4 @@ export default function sync () {
 	})
 
 	initialize.call(this)
-
-	return this
 }