+++ src/instance/noop.js
@@ -1,6 +1,6 @@
 export default {
 	noop: true,
-	destroy () { return this },
-	reveal () { return this },
-	sync () { return this },
+	destroy () {},
+	reveal () {},
+	sync () {},
 }