+++ test/instance/constructor.spec.js
@@ -69,12 +69,6 @@ describe('ScrollReveal', () => {
 			expect(result).to.exist
 			expect(result).to.be.a('function')
 		})
-
-		it('should support method chaining', () => {
-			const stub = sinon.stub(console, 'log')
-			expect(sr.destroy().reveal().sync().watch()).to.equal(sr)
-			stub.restore()
-		})
 	})
 
 	describe('Non-operational Instance', () => {
@@ -104,10 +98,6 @@ describe('ScrollReveal', () => {
 		it('should have a `sync` method', () => {
 			expect(noop.sync).to.exist
 			expect(noop.sync).to.be.a('function')
-		})
-
-		it('should support method chaining', () => {
-			expect(noop.destroy().reveal().sync().watch()).to.equal(noop)
 		})
 	})
 })