+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project tries its best to use [Semantic Versioning](http://semver.org/)
 - **Breaking:** The `reveal()` method ignores existing element configuration.
 - **Breaking:** The instance method `isSupported()` is now static.
 - **Breaking:** `options.distance` supports only `em` `px` and `%` values.
+- **Breaking:** ScrollReveal methods are no longer chainable.
 - **Breaking:** ScrollReveal requires a commercial license, unless for [GPL-3.0](https://opensource.org/licenses/GPL-3.0) compatible open source projects.
 - Elements in a reveal sequence are no longer grouped, and reveal progressively when visible.
 - ScrollReveal uses a single `matrix3d()` property, with the correct prefix and only when necessary. [#292](https://github.com/jlmakes/scrollreveal/issues/292)